Built-in itembank

A data frame with administrative information per item. Includes only items that are part of a Rasch model. See builtin_itemtable for an overview of all currently defined items.

builtin_itembank

Format

A data.frame with variables:

Name	Label

key	String indicating a specific Rasch model (the key)

item	Item name, gsed lexicon

tau	Difficulty estimate

label	Label (English)

instrument	Instrument code

domain	Domain code

mode	Administration mode

number	Item number

See also

Examples

head(builtin_itembank)
#>        key      item   tau
#> 1 gsed2206 aqicmc033 58.74
#> 2 gsed2206 aqicmc038 65.81
#> 3 gsed2206 aqicmc040 65.66
#> 4 gsed2206 aqicmc061 70.63
#> 5 gsed2206 aqifmc025 53.17
#> 6 gsed2206 aqifmc033 69.64
#>                                                                                                                                                                     label
#> 1                                                                                            "Does your child say 15 or more words in addition to ""Mama"" and ""Dada""?"
#> 2                                                                                  "When you ask, ""What is your name?"" does your child say her first name or nickname?"
#> 3                                                                               Does your child make sentences that are three or four words long? Please give an example:
#> 4                                                                               "When you ask, ""What is your name?"" does your child say both her first and last names?"
#> 5       Does your baby/child stack a small block or toy on top of another one? (You could also use spools of thread, small boxes, or toys that are about 1 inch in size.)
#> 6 After your child watches you draw a single circle, ask him to make a circle like yours. Do not let him trace your circle. Does your child copy you by drawing a circle?
#>   instrument domain mode number
#> 1        aqi     cm    c    033
#> 2        aqi     cm    c    038
#> 3        aqi     cm    c    040
#> 4        aqi     cm    c    061
#> 5        aqi     fm    c    025
#> 6        aqi     fm    c    033