Age-conditional reference distribution of D-score

A data frame containing the age-dependent distribution of the D-score for children aged 0-5 years. The distribution is modelled after the LMS distribution (Cole & Green, 1992), and is equal for both boys and girls. The LMS values can be used to graph reference charts and to calculate age-conditonal Z-scores, also known as DAZ.

builtin_references

Format

A data.frame with 265 rows and 17 variables:

Name	Label

pop	Population, either "dutch" or "gcdg"

age	Decimal age in years

mu	M-curve, median D-score, P50

sigma	S-curve, spread expressed as coefficient of variation

nu	L-curve, the lambda coefficient of the LMS model for skewness

P3	P3 percentile

P10	P10 percentile

P25	P25 percentile

P50	P50 percentile

P75	P75 percentile

P90	P90 percentile

P97	P97 percentile

SDM2	-2SD centile

SDM1	-1SD centile

SD0	0SD centile, median

SDP1	+1SD centile

SDP2	+2SD centile

Details

The "dutch" references were calculated from the SMOCC data, and cover age range 0-2.5 years (van Buuren, 2014). The "gcdg" references were calculated from the 15 cohorts of the GCDG-study, and cover age range 0-5 years (Weber, 2019).

References

Cole TJ, Green PJ (1992). Smoothing reference centile curves: The LMS method and penalized likelihood. Statistics in Medicine, 11(10), 1305-1319.

Van Buuren S (2014). Growth charts of human development. Stat Methods Med Res, 23(4), 346-368. pdf

Weber AM, Rubio-Codina M, Walker SP, van Buuren S, Eekhout I, Grantham-McGregor S, Caridad Araujo M, Chang SM, Fernald LCH, Hamadani JD, Hanlon A, Karam SM, Lozoff B, Ratsifandrihamanana L, Richter L, Black MM (2019). The D-score: a metric for interpreting the early development of infants and toddlers across global settings. BMJ Global Health, BMJ Global Health 4: e001724. pdf.

See also

Examples

head(builtin_references)

#>     pop    age    mu  sigma     nu   P3   P10   P25   P50   P75   P90   P97
#> 1 dutch 0.0383  8.81 0.3126 1.3917 3.15  5.07  6.91  8.81 10.57 12.07 13.49
#> 2 dutch 0.0575 10.59 0.2801 1.4418 4.32  6.49  8.52 10.59 12.50 14.12 15.64
#> 3 dutch 0.0767 12.27 0.2526 1.4891 5.61  7.96 10.10 12.27 14.28 15.97 17.56
#> 4 dutch 0.0958 13.87 0.2291 1.5331 6.99  9.43 11.64 13.87 15.93 17.67 19.30
#> 5 dutch 0.1150 15.39 0.2089 1.5722 8.42 10.89 13.13 15.39 17.47 19.23 20.89
#> 6 dutch 0.1342 16.83 0.1916 1.6049 9.86 12.32 14.56 16.83 18.92 20.69 22.36
#>   SDM2  SDM1   SD0  SDP1  SDP2
#> 1 2.78  5.94  8.81 11.39 13.76
#> 2 3.88  7.46 10.59 13.38 15.94
#> 3 5.12  8.98 12.27 15.20 17.87
#> 4 6.47 10.48 13.87 16.87 19.61
#> 5 7.89 11.95 15.39 18.43 21.21
#> 6 9.33 13.39 16.83 19.88 22.68