Sample of 50 children measured with three instruments

An example dataset with developmental scores at the item level for 50 random children from the GSED Validation Study (Cavellera et al, 2023). Each child has measurements from GSED SF (gs1), GSED LF (gl1) and BSID-III (by3).

triple

Format

A data.frame with 50 rows and 559 variables:

Name	Label

id	Integer, child ID

age	Numeric, age in decimal years

agedays	Integer, age in days

gs1sec001	Integer, SF001 Does your child smile?

gs1moc002	Integer, SF002 When lying on his/her back, ...

...	and so on..

The dataset contains 138 items from GSED SF (gs1), (item gs1moc028 was skipped), 155 items from GSED LF (gl1), and 263 (out of 326) items from BSID-III (by3).

References

Cavallera et al. (2023). Protocol for validation of the Global Scales for Early Development (GSED) for children under 3 years of age in seven countries. BMJ Open, 13(1), e062562. DOI: 10.1136/bmjopen-2022-062562. https://bmjopen.bmj.com/content/13/1/e062562

World Health Organization (WHO) (2023). Global Scales for Early Development (GSED) V1.0: Technical Report. Geneva: World Health Organization. https://www.who.int/publications/i/item/WHO-MSD-GSED-package-v1.0-2023.1

See also

Examples

# calculate D-score from all instruments
ds_all <- dscore(triple)
head(ds_all)
#>        a   n      p     d       sem    daz
#> 1 1.9493 114 0.6491 68.48 0.9153734  1.053
#> 2 2.5325  69 0.6522 71.73 1.0211466  0.165
#> 3 2.3874 101 0.5347 65.15 0.9611947 -1.145
#> 4 0.8980  32 0.5000 40.95 1.4663221 -1.646
#> 5 2.1903  72 0.3750 56.74 1.1374668 -2.501
#> 6 0.8980 121 0.7273 52.79 1.0302136  1.917
# calculate D-score from only GSED SF items
ds_sf <- dscore(triple, items = get_itemnames(instrument = "gs1"))
head(ds_sf)
#>        a  n      p     d      sem    daz
#> 1 1.9493 65 0.6769 68.95 1.210561  1.186
#> 2 2.5325 34 0.7059 73.23 1.458141  0.572
#> 3 2.3874 36 0.5833 65.89 1.404537 -0.966
#> 4 0.8980  8 0.5000 38.64 2.527097 -2.228
#> 5 2.1903 31 0.2258 57.84 1.605532 -2.289
#> 6 0.8980 80 0.7625 54.78 1.325298  2.517