Collection of items fitting the Rasch model

A data frame with administrative information per item with difficulty estimates (tau) from the Rasch model. The item bank provides the basic information to calculate D-scores. The items in the item bank are a subset of all items as collected in builtin_itemtable.

builtin_itembank

Format

A data.frame with variables:

Name	Label

key	String indicating a specific Rasch model

item	Item name, gsed lexicon

tau	Difficulty estimate

label	Label (English)

instrument	Instrument code

domain	Domain code

mode	Administration mode

number	Item number

Details

The difficulty estimates were estimated by a Rasch model. The key indicates the specific Rasch model used to estimate the difficulty. Strictly speaking, one can only compare D-score calculated from the same key.

Note

Updates:

Dec 01, 2022 - Overwrite labels of gto by correct item order.

Dec 05, 2022 - Adds key gsed2212, adding instruments gl1 and gs1, and defining correct order for gto

Jan 05, 2023 - Adds instrument gh1 to key gsed2212

See also

Examples

# count number of items per instrument in each key
table(builtin_itembank$instrument, builtin_itembank$key)
#>      
#>       293_0 dutch gcdg gsed1912 gsed2212 gsed2406
#>   aqi     0     0   29       17       17       17
#>   bar     0     0   15       13       13       13
#>   by1     0     0   85       76       76       76
#>   by2     0     0   16       16       16       16
#>   by3     0     0  105       67       67       67
#>   cro     0     0    0       62       64       64
#>   ddi     0    76   65       64       64       64
#>   den     0     0   67       50       50       50
#>   dmc     0     0    0       43       43       43
#>   ecd     0     0    0        0       18       18
#>   gh1     0     0    0        0       55       55
#>   gl1     0     0    0        0      155      155
#>   gpa   138     0    0        0      138      138
#>   gri     0     0  104       93       93       93
#>   gs1     0     0    0        0      139      139
#>   gsd     0     0    0        0        7        7
#>   gto   155     0    0        0      155      155
#>   iyo     0     0    0       55       57       57
#>   kdi     0     0    0       48       48       48
#>   mac     0     0    3        3        3        3
#>   mds     0     0    0        1        1        1
#>   mdt     0     0    0      126      126      126
#>   mul     0     0    0      138        0        0
#>   peg     0     0    1        1        1        1
#>   sbi     0     0    6        5        5        5
#>   sgr     0     0   19       19       19       19
#>   tep     0     0   33       31       31       31
#>   vin     0     0   17       17       17       17