Calculate posterior of ability

Calculate posterior of ability

calculate_posterior(scores, tau, qp, mu, sd)

Arguments

scores
A vector with PASS/FAIL observations. Scores are coded numerically as pass = 1 and fail = 0.

tau
A vector containing the item difficulties for the item scores in scores estimated from the Rasch model in the preferred metric/scale.

qp
Numeric vector of equally spaced quadrature points.

mu
Numeric scalar. The mean of the prior.

sd
Numeric scalar. Standard deviation of the prior.

Value

A list with three elements:

Name	Label

eap	Mean of the posterior

gp	Vcetor of quadrature points

posterior	Vector with posterior distribution.

Since dscore V40.1 the function does not return the "start" element.