Get labels for items

The get_labels() function obtains the item labels for a specified set of items.

get_labels(items = NULL, trim = NULL, itemtable = NULL)

Arguments

items

A character vector of item names to return. The default (NULL) returns the labels of all items.

trim

The maximum number of characters in the label. The default trim = NULL does not trim labels.

itemtable

A data.frame set up according to the same structure as builtin_itemtable(). If not specified, the builtin_itemtable is used.

Value

A named character vector with length(items) elements with item labels, in the same order as in items.

See also

Examples

# get labels of first two Macarthur items
get_labels(get_itemnames(instrument = "mac", number = 1:2), trim = 40)
#>                        macgmd001                        macgmd002 
#> "Child walks in a straight line"    "Walks backwards for 5 steps"